+++ mitra_activitypub/src/actors/mod.rs
@@ -1,4 +1,4 @@
 mod attachments;
 pub mod builders;
 pub(crate) mod handlers;
-pub(crate) mod keys;
+pub mod keys;