+++ mitra_api/src/activitypub/mod.rs
@@ -1,2 +1,3 @@
 mod receiver;
+mod types;
 pub mod views;