+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Make available space indication work on FreeBSD.
+- Resolve fragments when fetching objects.
 
 ### Changed