+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Accept poll options formatted as JSON object at `/api/v1/statuses`.
+- Require GET requests to portable inboxes and outboxes to be signed with DIDs.
 
 ## [4.0.0] - 2025-04-17